Fixing Asteroid Mats
Please do not just simply replace the Albeldo value on the prop
asteroid material (this messes it up for the entire game) you are not
creating a new material your simply changing the texture that is
already created) please ask how to create a new material or research it
online. You simply drag and drop the texture onto the gameObject in the
scene and unity will create the new material for you.