ShipList In Every Level!!! Fixing Material Mess ups. Joystick Image
For some reason GitHub messed up the game controller ship list in every
single level. So i went through each level and fixed the issue. I also
fixed respawning the player in level 4 (this was problematic due to the
difficulty of the level). I also fixed the mess ups on materials.
(Y’all have to create a new material not just change the texture) I
also changed the joystick image to Orange, and it changes to Red when
the ship is firing.